+++ packages/polaris-core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.51](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-core@2.0.0-beta.50...@enigmatis/polaris-core@2.0.0-beta.51) (2020-12-01)
+
+
+### Features
+
+* add dates filter ([#39](https://github.com/Enigmatis/polaris-united/issues/39)) ([705b639](https://github.com/Enigmatis/polaris-united/commit/705b639c6686af8a835f163c9ce2f2897e94d132))
+
+
+
+
+
 # [2.0.0-beta.50](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-core@2.0.0-beta.49...@enigmatis/polaris-core@2.0.0-beta.50) (2020-11-09)