+++ packages/polaris-graphql-logger/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/Enigmatis/polaris-graphql-logger/compare/@enigmatis/polaris-graphql-logger@2.2.2...@enigmatis/polaris-graphql-logger@2.3.0) (2020-12-01)
+
+
+### Features
+
+* add dates filter ([#39](https://github.com/Enigmatis/polaris-graphql-logger/issues/39)) ([705b639](https://github.com/Enigmatis/polaris-graphql-logger/commit/705b639c6686af8a835f163c9ce2f2897e94d132))
+
+
+
+
+
 ## [2.2.2](https://github.com/Enigmatis/polaris-graphql-logger/compare/@enigmatis/polaris-graphql-logger@2.2.1...@enigmatis/polaris-graphql-logger@2.2.2) (2020-11-09)